requirements.txt
Django==4.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
requests==2.28.2
coverage==7.2.3